script for activation

+++ cleaning.sh
@@ -1,4 +1,5 @@
-rm nohup.out
+#!/bin/bash
+[ -f nohup.out ] && rm nohup.out
 rm -rf ./parsl_logs/
 rm -rf /nfs_scratch/$LOGNAME/coffea_parsl_condor_htex/
-rm -rf /nfs_scratch/$LOGNAME/parsl_logs/runinfo/
+rm -rf /nfs_scratch/$LOGNAME/parsl_logs/
